+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2026-01-21
+
+### Added
+- `wtfb init-readme` command - Generate IMDb-style README from template
+  - Options: `--title`, `--type`, `--force`, `--dry-run`
+  - Supports project types: screenplay, novel, film-production
+  - Respects customized READMEs (skips unless `--force`)
+- IMDb-style README template
+
+### Changed
+- Updated Related links to point to story-systems-template repo
+
 ## [1.0.2] - 2026-01-11
 
 ### Added
@@ -32,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `wtfb export-html` - Export screenplay to HTML
 - `wtfb info` - Show CLI information
 
+[1.0.3]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/bybren-llc/wtfb-packages/releases/tag/v1.0.0